refactor(mercury): split channel/tests.rs (769 lines) into test directory
Split by subsystem:

- mod.rs: shared helpers + general tests
- reassembly.rs: fragment reassembly tests
- bundling.rs: packet bundling tests
