+++ crates/mercury/src/channel/tests/mod.rs
@@ -1,0 +1,7 @@
+//! Channel tests — split from a single tests.rs to keep each file
+//! under the 700-line cap. Submodules are grouped by subsystem.
+
+use super::*;
+
+mod channel_lifecycle;
+mod reassembly;
